Created new repo for alloys with precipitates
This set of scripts calculates yield locus for materials having precipitates. It has the option to consider precipitates as elastic or plastic inclusions. The elastic back stresses developed are calculated using Eshelby's inclusion method. For elastic inclusion approximation, none of the strain components are relaxed but for plastic inclusion approximation, some of the strain components of Eshelby's strain tensor are relaxed.